Semanal

+++ src/main/java/com/esteban/tienda/service/CategoriaService.java
@@ -5,10 +5,14 @@
 package com.esteban.tienda.service;
 import com.esteban.tienda.domain.Categoria;
 import com.esteban.tienda.repository.CategoriaRepository;
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+
 
 /**
  *
@@ -18,16 +22,23 @@ import org.springframework.transaction.annotation.Transactional;
 public class CategoriaService {
 
     private final CategoriaRepository categoriaRepository;
+    private final FirebaseStorageService firebaseStorageService;
 
-    public CategoriaService(CategoriaRepository categoriaRepository) {
+    public CategoriaService(
+            CategoriaRepository categoriaRepository,
+            FirebaseStorageService firebaseStorageService) {
+
         this.categoriaRepository = categoriaRepository;
+        this.firebaseStorageService = firebaseStorageService;
     }
 
     @Transactional(readOnly = true)
     public List<Categoria> getCategorias(boolean activo) {
+
         if (activo) {
             return categoriaRepository.findByActivoTrue();
         }
+
         return categoriaRepository.findAll();
     }
 
@@ -37,12 +48,52 @@ public class CategoriaService {
     }
 
     @Transactional
-    public void save(Categoria categoria) {
-        categoriaRepository.save(categoria);
+    public void save(Categoria categoria, MultipartFile imagenFile) {
+
+        categoria = categoriaRepository.save(categoria);
+
+        if (!imagenFile.isEmpty()) {
+
+            try {
+
+                String rutaImagen = firebaseStorageService.uploadImage(
+                        imagenFile,
+                        "categoria",
+                        categoria.getIdCategoria()
+                );
+
+                categoria.setRutaImagen(rutaImagen);
+
+                categoriaRepository.save(categoria);
+
+            } catch (IOException e) {
+
+            }
+        }
     }
 
     @Transactional
     public void delete(Integer idCategoria) {
-        categoriaRepository.deleteById(idCategoria);
+
+        if (!categoriaRepository.existsById(idCategoria)) {
+
+            throw new IllegalArgumentException(
+                    "La categoría con ID "
+                    + idCategoria
+                    + " no existe."
+            );
+        }
+
+        try {
+
+            categoriaRepository.deleteById(idCategoria);
+
+        } catch (DataIntegrityViolationException e) {
+
+            throw new IllegalStateException(
+                    "No se puede eliminar la categoria. Tiene datos asociados.",
+                    e
+            );
+        }
     }
 }
