+++ src/main/java/com/esteban/tienda/service/ProductoService.java
@@ -5,10 +5,13 @@
 package com.esteban.tienda.service;
 import com.esteban.tienda.domain.Producto;
 import com.esteban.tienda.repository.ProductoRepository;
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 /**
  *
@@ -18,16 +21,23 @@ import org.springframework.transaction.annotation.Transactional;
 public class ProductoService {
 
     private final ProductoRepository productoRepository;
+    private final FirebaseStorageService firebaseStorageService;
 
-    public ProductoService(ProductoRepository productoRepository) {
+    public ProductoService(
+            ProductoRepository productoRepository,
+            FirebaseStorageService firebaseStorageService) {
+
         this.productoRepository = productoRepository;
+        this.firebaseStorageService = firebaseStorageService;
     }
 
     @Transactional(readOnly = true)
     public List<Producto> getProductos(boolean activo) {
+
         if (activo) {
             return productoRepository.findByActivoTrue();
         }
+
         return productoRepository.findAll();
     }
 
@@ -37,36 +47,84 @@ public class ProductoService {
     }
 
     @Transactional
-    public void save(Producto producto) {
-        productoRepository.save(producto);
+    public void save(Producto producto, MultipartFile imagenFile) {
+
+        producto = productoRepository.save(producto);
+
+        if (!imagenFile.isEmpty()) {
+
+            try {
+
+                String rutaImagen = firebaseStorageService.uploadImage(
+                        imagenFile,
+                        "producto",
+                        producto.getIdProducto()
+                );
+
+                producto.setRutaImagen(rutaImagen);
+
+                productoRepository.save(producto);
+
+            } catch (IOException e) {
+
+            }
+        }
     }
 
     @Transactional
     public void delete(Integer idProducto) {
-        productoRepository.deleteById(idProducto);
+
+        if (!productoRepository.existsById(idProducto)) {
+
+            throw new IllegalArgumentException(
+                    "La categoría con ID "
+                    + idProducto
+                    + " no existe."
+            );
+        }
+
+        try {
+
+            productoRepository.deleteById(idProducto);
+
+        } catch (DataIntegrityViolationException e) {
+
+            throw new IllegalStateException(
+                    "No se puede eliminar la producto. Tiene datos asociados.",
+                    e
+            );
+        }
     }
 
     @Transactional(readOnly = true)
     public List<Producto> consultaDerivada(double precioInf, double precioSup) {
-        return productoRepository.findByPrecioBetweenOrderByPrecioAsc(precioInf, precioSup);
+        return productoRepository.findByPrecioBetweenOrderByPrecioAsc(
+                precioInf,
+                precioSup
+        );
     }
 
     @Transactional(readOnly = true)
     public List<Producto> consultaJPQL(double precioInf, double precioSup) {
-        return productoRepository.consultaJPQL(precioInf, precioSup);
+        return productoRepository.consultaJPQL(
+                precioInf,
+                precioSup
+        );
     }
 
     @Transactional(readOnly = true)
     public List<Producto> consultaSQL(double precioInf, double precioSup) {
-        return productoRepository.consultaSQL(precioInf, precioSup);
+        return productoRepository.consultaSQL(
+                precioInf,
+                precioSup
+        );
     }
-
     @Transactional(readOnly = true)
-    public List<Producto> consultaAmpliada(double precioInf,
-                                           double precioSup,
-                                           boolean activo) {
-        return productoRepository.findByPrecioBetweenAndActivoOrderByPrecioAsc(
-                precioInf, precioSup, activo);
-    }
-
+public List<Producto> consultaAmpliada(double precioInf, double precioSup, boolean activo) {
+    return productoRepository.findByPrecioBetweenAndActivoOrderByPrecioAsc(
+            precioInf,
+            precioSup,
+            activo
+    );
+}
 }
